Fixed a warning

+++ modules/juce_audio_processors/processors/juce_AudioProcessor.h
@@ -1426,8 +1426,8 @@ private:
     {
         Array<InOutChannelPair> layouts;
 
-        for (int i = 0; i < numLayouts; ++i)
-            layouts.add (InOutChannelPair (configuration[i]));
+        for (size_t i = 0; i < numLayouts; ++i)
+            layouts.add (InOutChannelPair (configuration[(int) i]));
 
         return layouts;
     }
